gocollect: Write core.id.regid with 0400 perms instead of 0644.
We expect to be run as root, and there is no need to disclose the file
contents to others on the system. We don't want them to be able to screw
with our data. And on top of that, we're supposed to write to it once,
and then leave it.