os.pkg: Lock /var/lib/{dpkg,apt/lists}/lock before doing os.pkg.
During an apt upgrade, the DB would be in flux, causing some packages to
be listed as (temporarily) deleted if gocollect ran during an apt
upgrade. When gocollect itself is upgraded, the chance of that happening
is significant.

The fix is to lock the lock files that apt/dpkg use before trying to
read the dpkg database. Unfortunately, this meant introducing a python
code helper. We could create a go or static C binary for this purpose,
but I think we can live with the optional python dependency for now.

If python does not exist, os.pkg will now fail on dpkg-based systems.