collectors: app.lshw: Make it optional.
We don't really need lshw, now that we have all the other nice
collectors.

This fixes an issue where on embedded systems the lshw package does not
exist in the repository.

We could put if in the Recommends list (if that gracefully succeeds if
it doesn't exist in the repository) but that's more work: we'd have to
add extra code for the Recommends line, and decide which others get to
go in there.

In the future we should try to make all app.* collectors optional:

    $ grep -c LABELS.*optional collectors/app.*
    collectors/app.dmidecode:0
    collectors/app.lshw:1
    collectors/app.ps-kvmex1:1
    collectors/app.vzlist:1
    collectors/app.vzlistex1:1