os.storage: Reduce os.storage listing.
From now on, only the /devices are listed, no "magicfs" like tmpfs or
zfs. This is because they -- at least ZFS on LXC guests -- tend to
fluctuate.

We add a ``zpool list`` to the output so we don't lose the ZFS entry
altogether, but we'll lose all ZFS containers (which generally share the
same large disk pool anyway, and which can fluctuate if you spawn
containers at will).

Closes #14.