golang: No need to use pointers when using interfaces.
os.FileInfo is an interface, so the data was already passed by
reference.

Taking the following example:

    type I interface {
      SetN(string)
    }
    type U struct {
      Name string
    }
    func main() {
      u := U{Name: "Leto"}
      println(u.Name)
      Modify(&u)
      println(u.Name)
    }
    func Modify(i I) {
      i.SetN("Paul")
    }

    // This yields: "Leto", "Leto"
    func (u U) SetN(n string) {
      u.Name = n
    }
    // This yields: "Leto", "Paul"
    func (u *U) SetN(n string) {
      u.Name = n
    }

The interface already acts as if a pointer was passed (but only if SetN
has a pointer receiver (the 2nd SetN)). Assuming that that's the case,
we don't need the extra pointers. If it isn't the case, compile errors
happen elsewhere.

Before this changeset, we were rather needlessly passing pointers to
pointers.