os.keys: Fix encoding in apt-key uid listing
If nothing is specified (LANG=C?), it apparently spits out latin1. Not
so nice, as this wans't apparent during tests, and it caused:

    UnicodeDecodeError: 'utf-8' codec can't decode byte 0xe4 in position
      4037: invalid continuation byte

in:

    {"algo":"rsa2048","created":"2011-08-03","expires":"2020-07-25",
     "key":"30EBF4E73CCE63EEE124DD278E6DA8B4E158C569",
     "uid":"... Sepp\xe4nen (Op..."},

Note that:

(a) gocollect (go) did not choke on the invalid UTF-8
(b) the gocollect-endpoint did

Ideally gocollect-client should not send invalid UTF-8, and just replace
invalid characters. And the endpoint should perhaps not choke on it
either, but replace and warn.