rc: Depend on networking before start
This should fix issues where gocollect is started before networkin is
up, and then fails to push immediately after boot.

We generally want this post-boot push, so having it fail is bad.

For upstart we did not change anything, as upstart is unused and it
sucks and there is no sane way to configure this.