init: Try to run last, after the system has booted
Yes, it looks conflicting, to be both on WantedBy= and After=. But the
After= should define ordering if there is any. The WantedBy= should
place it in the to-be-run-list.

The intention is that this runs after the system is fully booted
(whether in running or degraded state does not matter). This way we can
expect few changes (because things like mounts, modules, sysctl, cpu
settings have all been set). Ideally gocollect collectors have 0 changes
between reboots (except for the uptime).